# docs/openapi.yaml
openapi: 3.0.3
info:
  title: My Interview API
  version: 1.0.0
  description: Frontend contract verified against the current Spring Boot implementation.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/interview-sessions:
    post:
      tags: [Interview]
      operationId: createInterviewSession
      summary: Create and prepare an interview session
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateInterviewSessionRequest'}
      responses:
        '202':
          description: Existing or newly created session; preparation is asynchronous.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewSessionStatus'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/interview-sessions/{sessionId}:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    get:
      tags: [Interview]
      operationId: getInterviewSession
      summary: Get interview session status
      responses:
        '200':
          description: Session state.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewSessionStatus'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/interview-sessions/{sessionId}/preparation/retry:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    post:
      tags: [Interview]
      operationId: retryInterviewPreparation
      summary: Retry failed interview preparation
      responses:
        '202':
          description: Preparation redispatched.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewSessionStatus'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /api/interview-sessions/{sessionId}/start:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    post:
      tags: [Interview]
      operationId: startInterview
      summary: Start a prepared interview
      responses:
        '200':
          description: Started or already-running conversation.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewConversation'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
  /api/interview-sessions/{sessionId}/conversation:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    get:
      tags: [Interview]
      operationId: getInterviewConversation
      summary: Get the persisted conversation
      responses:
        '200':
          description: Persisted conversation and timer snapshot.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewConversation'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/interview-sessions/{sessionId}/answers:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    post:
      tags: [Interview]
      operationId: submitInterviewAnswer
      summary: Submit an answer and generate a reply
      description: Long-running synchronous AI request; reuse the same key and payload on retry.
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SubmitInterviewAnswerRequest'}
      responses:
        '200':
          description: Candidate turn plus persisted interviewer reply, if one was produced.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewAnswerResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '500': {$ref: '#/components/responses/ServerError'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
        '504': {$ref: '#/components/responses/GatewayTimeout'}
...
  /api/interview-sessions/{sessionId}/finish:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    post:
      tags: [Interview]
      operationId: finishInterview
      summary: Finish an interview early
      responses:
        '200':
          description: Conversation after moving to scoring.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InterviewConversation'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
  /api/admin/interview-sessions/{sessionId}:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    get:
      tags: [Admin]
      operationId: getAdminInterviewSession
      summary: Get interview-session operational details
      responses:
        '200':
          description: Session metadata, processing errors, and transition history.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminSessionDetail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/admin/interview-sessions/{sessionId}/preparation/retry:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    post:
      tags: [Admin]
      operationId: retryAdminInterviewPreparation
      summary: Retry failed interview preparation as an administrator
      responses:
        '202':
          description: Preparation redispatched.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminSessionDetail'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  parameters:
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 100}
...
  responses:
    BadRequest:
      description: Invalid input.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Unauthorized:
      description: Missing or invalid authentication.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Forbidden:
      description: Authenticated but not allowed.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    NotFound:
      description: Resource not found or inaccessible.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Conflict:
      description: State, idempotency, version, or limit conflict.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
...
    ServerError:
      description: Internal or AI-output error.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    ServiceUnavailable:
      description: AI, storage, or configuration unavailable.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    GatewayTimeout:
      description: AI or speech provider timed out.
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
...
  schemas:
    ApiError:
      type: object
      required: [timestamp, status, code, message, path]
      properties:
        timestamp: {type: string, format: date-time}
        status: {type: integer}
        code: {type: string}
        message: {type: string}
        path: {type: string}
        fieldErrors:
          type: object
          nullable: true
          additionalProperties: {type: string}
...
    AdminUserReference:
      type: object
      required: [id, fullName, email]
      properties:
        id: {type: integer, format: int64}
        fullName: {type: string}
        email: {type: string, format: email}
...
    InterviewTransitionActor:
      type: string
      enum: [USER, ADMIN, SYSTEM, SCHEDULER]
    AdminSessionTransition:
      type: object
      required: [id, toStatus, actor, occurredAt]
      properties:
        id: {type: integer, format: int64}
        fromStatus:
          allOf: [{$ref: '#/components/schemas/InterviewSessionStatusEnum'}]
          nullable: true
        toStatus: {$ref: '#/components/schemas/InterviewSessionStatusEnum'}
        reason: {type: string, nullable: true}
        actor: {$ref: '#/components/schemas/InterviewTransitionActor'}
        occurredAt: {type: string, format: date-time}
    AdminSessionDetail:
      type: object
      required: [id, user, status, templateTitle, profileName, languageCode, durationMinutes, interviewerStyle, mode, currentTurnIndex, createdAt, updatedAt, transitions]
      properties:
        id: {type: integer, format: int64}
        user: {$ref: '#/components/schemas/AdminUserReference'}
        status: {$ref: '#/components/schemas/InterviewSessionStatusEnum'}
        templateTitle: {type: string}
        profileName: {type: string}
        languageCode: {type: string}
        durationMinutes: {type: integer, minimum: 1}
        interviewerStyle: {$ref: '#/components/schemas/InterviewerStyle'}
        mode: {$ref: '#/components/schemas/InterviewSessionMode'}
        realtimeProvider: {type: string, nullable: true}
        realtimeVoiceName: {type: string, nullable: true}
        currentTurnIndex: {type: integer, minimum: 0}
        preparationErrorCode: {type: string, nullable: true}
        preparationErrorMessage: {type: string, nullable: true}
        planSchemaVersion: {type: string, nullable: true}
        planModelName: {type: string, nullable: true}
        planPromptVersion: {type: string, nullable: true}
        scoringErrorCode: {type: string, nullable: true}
        scoringErrorMessage: {type: string, nullable: true}
        preparationStartedAt: {type: string, format: date-time, nullable: true}
        preparedAt: {type: string, format: date-time, nullable: true}
        startedAt: {type: string, format: date-time, nullable: true}
        deadlineAt: {type: string, format: date-time, nullable: true}
        lastActivityAt: {type: string, format: date-time, nullable: true}
        endReason: {$ref: '#/components/schemas/InterviewEndReason'}
        endedAt: {type: string, format: date-time, nullable: true}
        scoringStartedAt: {type: string, format: date-time, nullable: true}
        completedAt: {type: string, format: date-time, nullable: true}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
        transitions:
          type: array
          items: {$ref: '#/components/schemas/AdminSessionTransition'}
...
    InterviewerStyle: {type: string, enum: [FRIENDLY, PROFESSIONAL, CHALLENGING]}
    InterviewSessionMode: {type: string, enum: [TURN_BASED, VOICE_REALTIME]}
    InterviewTurnInputMode: {type: string, enum: [TEXT, VOICE, VOICE_REALTIME]}
    InterviewSessionStatusEnum:
      type: string
      enum: [PREPARING, READY, PREPARATION_FAILED, IN_PROGRESS, SCORING, COMPLETED, SCORING_FAILED, CANCELLED, EXPIRED]
    InterviewEndReason:
      type: string
      enum: [AI_COMPLETED, TIME_EXPIRED, CANDIDATE_FINISHED, SYSTEM_TERMINATED]
      nullable: true
...
    CreateInterviewSessionRequest:
      type: object
      required: [templateId, profileId, languageCode, durationMinutes, interviewerStyle]
      properties:
        templateId: {type: integer, format: int64, minimum: 1}
        profileId: {type: integer, format: int64, minimum: 1}
        languageCode: {type: string, minLength: 1, maxLength: 10}
        durationMinutes: {type: integer, minimum: 1}
        interviewerStyle: {$ref: '#/components/schemas/InterviewerStyle'}
        mode: {$ref: '#/components/schemas/InterviewSessionMode'}
    InterviewSessionStatus:
      type: object
      required: [id, status, templateTitle, profileName, languageCode, durationMinutes, interviewerStyle, mode]
      properties:
        id: {type: integer, format: int64}
        status: {$ref: '#/components/schemas/InterviewSessionStatusEnum'}
        templateTitle: {type: string}
        profileName: {type: string}
        languageCode: {type: string}
        durationMinutes: {type: integer}
        interviewerStyle: {$ref: '#/components/schemas/InterviewerStyle'}
        mode: {$ref: '#/components/schemas/InterviewSessionMode'}
        realtimeProvider: {type: string, nullable: true}
        realtimeVoiceName: {type: string, nullable: true}
        preparationErrorCode: {type: string, nullable: true}
        preparationErrorMessage: {type: string, nullable: true}
        scoringErrorCode: {type: string, nullable: true}
        scoringErrorMessage: {type: string, nullable: true}
        preparedAt: {type: string, format: date-time, nullable: true}
        endReason: {$ref: '#/components/schemas/InterviewEndReason'}
        endedAt: {type: string, format: date-time, nullable: true}
        completedAt: {type: string, format: date-time, nullable: true}
...
    CandidateIntent:
      type: string
      enum: [ANSWER, REQUEST_REPEAT, REQUEST_CLARIFICATION, REQUEST_TIME, ASK_INTERVIEWER, CANNOT_ANSWER, DECLINE_OR_SKIP, CORRECT_PREVIOUS_ANSWER, REQUEST_END, SOCIAL_OR_META, OFF_TOPIC, INAPPROPRIATE, OTHER]
      nullable: true
    InterviewTurnAction:
      type: string
      enum: [OPENING, EXPLORE, FOLLOW_UP, HANDLE_REQUEST, CLOSE]
      nullable: true
    InterviewTurn:
      type: object
      required: [id, turnIndex, role, inputMode, content, wasInterrupted, createdAt]
      properties:
        id: {type: integer, format: int64}
        turnIndex: {type: integer}
        role: {type: string, enum: [INTERVIEWER, CANDIDATE]}
        inputMode: {$ref: '#/components/schemas/InterviewTurnInputMode'}
        content: {type: string}
        candidateIntent: {$ref: '#/components/schemas/CandidateIntent'}
        action: {$ref: '#/components/schemas/InterviewTurnAction'}
        focusAreaCode: {type: string, nullable: true}
        requestId: {type: string, nullable: true}
        processingStatus: {type: string, enum: [PROCESSING, COMPLETED, FAILED], nullable: true}
        processingErrorCode: {type: string, nullable: true}
        wasInterrupted: {type: boolean}
        latencyMs: {type: integer, minimum: 0, nullable: true}
        createdAt: {type: string, format: date-time}
    InterviewConversation:
      type: object
      required: [sessionId, status, mode, remainingSeconds, currentTurnIndex, turns]
      properties:
        sessionId: {type: integer, format: int64}
        status: {$ref: '#/components/schemas/InterviewSessionStatusEnum'}
        mode: {$ref: '#/components/schemas/InterviewSessionMode'}
        realtimeProvider: {type: string, nullable: true}
        realtimeVoiceName: {type: string, nullable: true}
        startedAt: {type: string, format: date-time, nullable: true}
        deadlineAt: {type: string, format: date-time, nullable: true}
        endReason: {$ref: '#/components/schemas/InterviewEndReason'}
        endedAt: {type: string, format: date-time, nullable: true}
        remainingSeconds: {type: integer, format: int64, minimum: 0}
        currentTurnIndex: {type: integer, minimum: 0}
        turns: {type: array, items: {$ref: '#/components/schemas/InterviewTurn'}}
    SubmitInterviewAnswerRequest:
      type: object
      required: [expectedTurnIndex, answer, inputMode]
      properties:
        expectedTurnIndex: {type: integer, minimum: 0}
        answer: {type: string, minLength: 1, maxLength: 8000}
        inputMode: {type: string, enum: [TEXT, VOICE]}
...
    InterviewAnswerResponse:
      type: object
      required: [sessionId, status, remainingSeconds, currentTurnIndex]
      properties:
        sessionId: {type: integer, format: int64}
        status: {$ref: '#/components/schemas/InterviewSessionStatusEnum'}
        deadlineAt: {type: string, format: date-time, nullable: true}
        endReason: {$ref: '#/components/schemas/InterviewEndReason'}
        endedAt: {type: string, format: date-time, nullable: true}
        remainingSeconds: {type: integer, format: int64, minimum: 0}
        currentTurnIndex: {type: integer, minimum: 0}
        candidateTurn:
          type: object
          allOf: [{$ref: '#/components/schemas/InterviewTurn'}]
          nullable: true
        interviewerTurn:
          type: object
          allOf: [{$ref: '#/components/schemas/InterviewTurn'}]
          nullable: true
...
security:
  - bearerAuth: []